+++ SoftwareEngineerSkills/src/SoftwareEngineerSkills.Domain/Common/Events/DomainEvent.cs
@@ -3,7 +3,7 @@ namespace SoftwareEngineerSkills.Domain.Common.Events;
 /// <summary>
 /// Base class for domain events
 /// </summary>
-public abstract class DomainEvent
+public abstract class DomainEvent : IDomainEvent
 {
     /// <summary>
     /// Gets the date and time when the event occurred
@@ -13,14 +13,24 @@ public abstract class DomainEvent
     /// <summary>
     /// Gets a unique identifier for the event instance
     /// </summary>
-    public Guid EventId { get; }
+    public Guid Id { get; }
     
     /// <summary>
     /// Creates a new instance of a domain event
     /// </summary>
     protected DomainEvent()
     {
-        EventId = Guid.NewGuid();
+        Id = Guid.NewGuid();
         OccurredOn = DateTime.UtcNow;
+    }
+    
+    /// <summary>
+    /// Creates a new instance of a domain event with a specified occurrence time
+    /// </summary>
+    /// <param name="occurredOn">The date and time when the event occurred</param>
+    protected DomainEvent(DateTime occurredOn)
+    {
+        Id = Guid.NewGuid();
+        OccurredOn = occurredOn;
     }
 }