+++ SoftwareEngineerSkills/tests/SoftwareEngineerSkills.Infrastructure.UnitTests/Persistence/Repositories/EfRepositoryTests.cs
@@ -11,6 +11,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore.Query;
 using Xunit;
 
 namespace SoftwareEngineerSkills.Infrastructure.UnitTests.Persistence.Repositories;